Add Kotlin compiler settings and new solution file
A new Kotlin compiler settings file has been added in the .idea directory, to standardize the compiler version to JVM 1.8 and Kotlin language version to 1.9. Additionally, a new Kotlin class (Solution) has been added to calculate and print the day of the week for a given date, located within a new 'DayOfTheWeek' directory.